Time: 11 ms (42.21%), Space: 18.4 MB (71.03%) - LeetHub

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -33,7 +33,7 @@ public:
             if(num==num1)c1++;
             if(num==num2)c2++;
         }
-        int mini=n/3+1;
+        int mini=int(n/3)+1;
         if(c1>=mini)
             ans.push_back(num1);
          if(c2>=mini)
